% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_estimateR.R
\name{model_estimateR}
\alias{model_estimateR}
\title{This function is used to estimate  the  different paramters of interest like the time varying transmission rates, proportion of reported cases and thebasic reproduction rates.}
\usage{
model_estimateR(
  data,
  data_init,
  N,
  init_pars = NULL,
  niter = 1e+05,
  BurnIn = 5000,
  model = "Multinomial",
  plot = TRUE,
  period_start,
  auto.initialize = TRUE,
  ...
)
}
\arguments{
\item{data}{(Mandatory) input :If the model is Multinomial then the data matrix should contain 3 columns namely Confirmed,Recovered and Death, while if the model is Poisson or Binomial, then the data should contain only one column namely Confirmed. Please see that the names of the colums are maintained.}

\item{data_init}{(Mandatory) This is the initial data values that should be provided by the user which is a six size length vector. The first entry should contain the Total Confirmed, Total Recovered ,Total Death , Daily Confirmed ,Daily Recovered and Daily Death for the Starting Date.If the starting total confirmed is 0 , please replace it by 1}

\item{N}{(Mandatory) : The population size.}

\item{init_pars}{= NULL( default) . Else the user can give a user input initial paramters which should consist of the initial values of the time varying beta , proportion of testing for the different periods.}

\item{niter}{= 1e5 (default)  No of iteration for the MCMC metropolis hasting algorithm.}

\item{BurnIn}{= 5e4 (default) This is the Burn-In Period for the MCMC algorithm}

\item{model}{= Multinomial(DEFAULT) This is the likelihood function that will be used. There are three options available namely Multinomial, Poisson and Binomial.}

\item{plot}{= TRUE(Default) This will give the box plot for the basic reproductuon number for the different periods.}

\item{period_start}{The total time period is divided into small periods depending on the lockdowm measures imposed by the goverment. So this is a vector consisting of the start dates for the different time periods.}

\item{auto.initialize}{= TRUE(default) This is the option for using a mle based initial parameter.}
}
\value{

}
\description{
This function is used to estimate  the  different paramters of interest like the time varying transmission rates, proportion of reported cases and thebasic reproduction rates.
}
