% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.SEIRfansy}
\alias{plot.SEIRfansy}
\title{Plot an SEIRfansy Fit Object}
\usage{
\method{plot}{SEIRfansy}(x, type, ...)
}
\arguments{
\item{x}{SEIRfansy object to plot}

\item{type}{type of plot to render. Options are "trace" or "boxplot".}

\item{...}{not currently used}
}
\description{
This is a convenient wrapper for output that is already included in the
SEIRfansy output in the plots element. Options are trace or boxplot.
}
